hooks: extract useSetFlash to its own file
